refactor(dto): Score 객체를 dto로 변경
