+++ src/main/java/baseball/domain/dto/Score.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.dto;
 
 import java.util.ArrayList;
 import java.util.Arrays;
